test files are moved
