+++ components/tests/components/NameForm.test.tsx
@@ -6,7 +6,7 @@ import {
   waitForElementToBeRemoved,
 } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { NameForm } from '../Screens/Name/NameForm';
+import { NameForm } from '../../Screens/Name/NameForm';
 
 describe('NameForm', () => {
   const props = {
