+++ setupTests.js
@@ -1,0 +1,1 @@
+process.env.END_POINT_URL = 'http://localhost:3000';
